ml/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.0
filelock==3.10.7
future==0.18.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
pydantic==1.10.7
python-whois==0.8.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
torch==2.0.0
typing_extensions==4.5.0
uvicorn==0.21.1
